+++ constants/theme.js
@@ -1,0 +1,6 @@
+export default {
+  black: '#0E1014',
+  white: '#ffffff',
+
+  primary: '#CDABFD',
+};
