+++ package.json
@@ -13,7 +13,6 @@
     "@react-navigation/stack": "^5.9.0",
     "axios": "^0.20.0",
     "expo": "~38.0.8",
-    "expo-linear-gradient": "~8.2.1",
     "expo-status-bar": "^1.0.2",
     "react": "~16.11.0",
     "react-dom": "~16.11.0",
